moved get_args() function into args.py, attempting to add subparsers
