+++ peek-app/functions.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding:utf-8 -*-
 
-import argparse
 import sys
 
 # remove traceback from errors
@@ -93,71 +92,3 @@ def tail(*args) -> list:
         return output
     except IndexError as error:
         return error
-
-
-def get_args():
-    """
-    Parse console args and return three variables
-    that contain commands, flags, and additional args
-
-    Values:
-        None:           Accepts inputs and returns values
-
-    Return (list):      Three variables that contains the path to dataset,
-                        flags, fields (if any)
-    """
-
-    # create parser object
-    parser = argparse.ArgumentParser()
-
-    # arg for filepath
-    parser.add_argument(
-        "-f", "--filepath",
-        required=True,
-        dest="filepath",
-        help="filepath of dataset")
-
-    # func for search
-    parser.add_argument(
-        "-s", "--search",
-        nargs=2,
-        dest="search",
-        default=False,
-        help="Search for matching results based on field and value")
-
-    # func for listing fields
-    parser.add_argument(
-        "-l", "--list",
-        dest="list",
-        action="store_true",
-        default=False,
-        help="Returns list of fields available for filtration"
-    )
-
-    # func for head of data
-    parser.add_argument(
-        "-hd", "--head",
-        dest="head",
-        action="store_true",
-        default=False,
-        help="Shows preview of first 5 rows of dataset"
-    )
-
-    # func for tail of data
-    parser.add_argument(
-        "-t", "--tail",
-        dest="tail",
-        action="store_true",
-        default=False,
-        help="Shows preview of last 5 rows of dataset"
-    )
-
-    # gather arguments and flags into variables
-    args = parser.parse_args()
-    data = args.filepath
-    commands = [tup for tup in list(
-        vars(args).items())[1:] if tup[1] is not False][0]
-    flags = commands[0]
-    fields = commands[1]
-
-    return data, commands, flags, fields
